fixing code review comments
